# Brain_API_Specification.yaml
openapi: 3.1.0
info:
  title: Brain API
  version: 0.3.0
  description: |
    The Brain API exposes six layers as HTTP endpoints plus one MCP surface.
    Raw (evidence), Ledger (normalized financial truth), Wiki (memory),
    Policy (governance), Agent (action orchestration), Audit (proof).

    v0.2.0 introduces the Normalized Ledger layer. The Wiki API is preserved
    but its role narrows to memory rendering and Q&A — financial truth lives
    in the Ledger. The Execution endpoints are renamed to Agent endpoints;
    legacy /execution/* paths remain shipped with deprecation headers for
    the duration of the v0.3 transition.

    v0.3.0 aligns the spec to the canonical public docs at
    https://docs.brain.fi. Material changes land in small commits — see
    docs/sdk-audit.md for the reconciliation queue. This release aligns
    the error envelope to the docs shape (`{ error: { code, message,
    details, trace_id, docs_url } }`); the v0.2 flat shape is replaced.

    All endpoints are tenant-scoped. Tenant identity is derived from the
    authenticated principal. The principal's tenant is never passed in URLs
    except where explicitly required for cross-tenant operations by platform
    administrators.

    Base URL (production): https://api.brain.fi/v1
    Base URL (sandbox):    https://api.brain.dev/v1
    Protocol: HTTPS only. HTTP requests are rejected at the edge.
    Content type: application/json unless otherwise specified.
    Character encoding: UTF-8.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /actions:
    post:
      tags: [Action]
      summary: Propose an action
      description: |
        Creates an Action (internally a PaymentIntent row) and attaches
        the PolicyDecision. On `decision === "ALLOW"` the action is
        auto-approved and ready for `/execute`; on `"ESCALATE"` it
        waits in `status === "needs_approval"`; on `"DENY"` it is
        terminal in `status === "rejected"`.
      operationId: createAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenantId, type]
              properties:
                tenantId: { type: string }
                agent_id:
                  type: string
                  description: |
                    Proposing agent. Required when the caller is an
                    agent principal; the server falls back to the
                    auth context if omitted.
                type:
                  type: string
                  description: |
                    Free-form action type (e.g. `pay_invoice`,
                    `outbound_payment`, `onchain_transfer`,
                    `erp_writeback`, `card_payment`). The policy DSL
                    `applies_to` clause gates each value.
                # Convenience shorthand: caller passes either `invoiceId`
                # OR an explicit destination + amount.
                invoiceId:
                  type: string
                  description: |
                    When set, the server populates source account /
                    destination / amount / currency from the linked
                    ledger_invoices row.
                to:
                  type: object
                  properties:
                    counterparty_id: { type: string }
                amount:
                  type: string
                  pattern: '^\d+(\.\d+)?$'
                currency:
                  type: string
                  pattern: '^[A-Z]{3}$'
                source_account_id: { type: string }
                memo:
                  type: string
                  maxLength: 500
                evidence_ids:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Action created with policy decision attached
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Action' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }

    get:
      tags: [Action]
      summary: List actions
      operationId: listActions
      parameters:
        - name: tenantId
          in: query
          schema: { type: string }
        - name: agent_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ActionStatus' }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Action list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Action' }
                  next_cursor:
                    type: string
                    nullable: true
  /actions/{action_id}:
    get:
      tags: [Action]
      summary: Get a single action with its PolicyDecision + audit trail
      operationId: getAction
      parameters:
        - name: action_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Action
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Action' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Action]
      summary: Cancel a proposed action
      description: |
        Only Actions in `status === "needs_approval"` (or pre-approval
        `auto`) can be cancelled. Cancellation is a terminal state.
      operationId: cancelAction
      parameters:
        - name: action_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cancelled action
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Action' }
        '409':
          description: Action not in a cancellable state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /actions/{action_id}/approve:
    post:
      tags: [Action]
      summary: Sign approval on an ESCALATE action
      operationId: approveAction
      parameters:
        - name: action_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                as:
                  type: string
                  description: |
                    Approver identity reference (e.g. `user_cfo`,
                    `role:treasurer`). Default: the caller's principal id.
                approver_role:
                  type: string
                  description: Role context for the approval.
                signature:
                  type: string
                  description: |
                    Hex-encoded EIP-712 signature when the policy is
                    on-chain-registered. Optional for off-chain policies.
      responses:
        '200':
          description: Updated action (approval recorded; status may transition to `approved`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Action' }
  /actions/{action_id}/reject:
    post:
      tags: [Action]
      summary: Reject an action
      operationId: rejectAction
      parameters:
        - name: action_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                as: { type: string }
                reason: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Rejected action
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Action' }
  /actions/{action_id}/execute:
    post:
      tags: [Action]
      summary: Execute an approved action through its rail
      description: |
        Runs the deterministic 13-step pre-execution gate (§6 of
        Engineering Standards) before dispatching to the rail. On any
        gate failure returns 422 with one of the `gate_*` error codes
        (see https://docs.brain.fi/resources/errors).
      operationId: executeAction
      parameters:
        - name: action_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Dispatched to rail
          content:
            application/json:
              schema:
                type: object
                properties:
                  action_id: { type: string }
                  execution_id: { type: string }
                  rail: { $ref: '#/components/schemas/ActionRail' }
                  status:
                    type: string
                    enum: [dispatched, in_flight]
        '422':
          description: Pre-execution gate failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  #############################################################################
  # v0.2 PAYMENT INTENTS (deprecated — use /actions/* above)
  #############################################################################
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token issued by Brain's auth service.
        Contains tenant_id and scopes. See Engineering Standards
        section 2 for the full auth model.
...
  responses:
    BadRequest:
      description: Request is malformed or missing required fields
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Authenticated but not authorized for this operation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found or not accessible to this tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: |
        Canonical Brain error envelope. Every non-2xx response wraps its
        body in this shape. Source of truth:
        https://docs.brain.fi/resources/errors (canonical wire codes) and
        https://docs.brain.fi/api-reference/overview (envelope shape).

        Engineering Standards §4.1 previously specified the same nested
        shape under the field name `request_id`; the docs publish it as
        `trace_id` and the docs win — see docs/sdk-audit.md conflict E.
      properties:
        error:
          type: object
          required: [code, message, trace_id, docs_url]
          properties:
            code:
              type: string
              description: |
                Machine-readable error code. Stable: once shipped a code
                is never renamed (Standards §4.3 "additions only"). See
                services/api/src/shared/errors.ts for the registry.
            message:
              type: string
              description: Human-readable error message (English, untranslated).
            details:
              type: object
              additionalProperties: true
              description: |
                Optional structured payload. Schema is code-specific —
                e.g. `policy_denied` carries `{rule_id, …}`, `gate_*`
                carries `{check, …}`.
            trace_id:
              type: string
              description: |
                Correlation id for support. Per-request, surfaced in
                logs, metrics, and traces. Pass to support to retrieve
                the full request timeline.
              example: trc_8f3a92...
            docs_url:
              type: string
              format: uri
              description: |
                Direct link to the docs page for this error code, of the
                form https://docs.brain.fi/errors/{code}.
              example: https://docs.brain.fi/errors/policy_denied
      example:
        error:
          code: policy_denied
          message: Counterparty not approved
          details:
            counterparty_id: cp_x
            policy_version: 3
          trace_id: trc_8f3a92abc
          docs_url: https://docs.brain.fi/errors/policy_denied

    #
    # Sources (PLAN-FIRST #12) — adapter connections that feed the Raw
    # layer. The 8-value MVP set per docs/sdk-audit.md decision K2.
    # Adding a new value to `SourceType` is a non-breaking enum widening
    # — extends as adapters ship (Architecture §3.1, "Other
    # source-specific adapters … are post-MVP").
    #
...
    ActionStatus:
      type: string
      description: |
        Caller-facing lifecycle state, per
        https://docs.brain.fi/api-reference/actions-api. Maps onto the
        internal PaymentIntent status one-to-one but uses the
        docs-publishing vocabulary instead of the storage one.
      enum:
        - auto
        - needs_approval
        - approved
        - rejected
        - executed
        - failed
        - cancelled
    ActionRail:
      type: string
      description: |
        Execution rail. Source:
        https://docs.brain.fi/api-reference/actions-api.
      enum:
        - bank_api
        - smart_account
        - x402
    Action:
      description: |
        v0.3 user-facing Action object — the resource served by
        `/v1/actions/*`. Internally backed by a PaymentIntent row in
        the Ledger; the route layer translates the storage enum
        vocabularies to the docs vocabulary on every response.

        Source: https://docs.brain.fi/api-reference/actions-api.
      type: object
      required: [id, tenantId, type, status, decision, expires_at]
      properties:
        id:
          type: string
          description: Action id (action_<ulid>).
        tenantId:
          type: string
          description: Owning tenant.
        agent_id:
          type: string
          nullable: true
          description: Proposing agent (null when proposed by a human).
        type:
          type: string
          description: |
            Action type. Examples: `pay_invoice`, `outbound_payment`,
            `onchain_transfer`, `erp_writeback`. Free-form string —
            the policy DSL `applies_to` clause is what gates each value.
        decision:
          type: string
          enum: [ALLOW, ESCALATE, DENY]
          description: |
            Latest policy verdict on this action. ALLOW = no approval
            needed; ESCALATE = waiting on signers in `approvers`;
            DENY = blocked. Matches `PolicyDecision.decision`.
        approvers:
          type: array
          description: |
            Approver references (e.g. `["role:cfo"]`). Populated when
            `decision === "ESCALATE"`.
          items: { type: string }
        approvals:
          type: array
          description: Approvals submitted so far.
          items:
            type: object
            properties:
              signer: { type: string }
              signed_at: { type: string, format: date-time }
              signature: { type: string }
        status:
          $ref: '#/components/schemas/ActionStatus'
        rail:
          $ref: '#/components/schemas/ActionRail'
        tx_hash:
          type: string
          nullable: true
          description: On-chain tx hash if `rail === "smart_account"`; else null.
        rail_receipt:
          type: object
          nullable: true
          description: |
            Rail-specific receipt body (ACH return file row, on-chain
            tx receipt, etc.). Schema is rail-dependent.
          additionalProperties: true
        executed_at:
          type: string
          format: date-time
          nullable: true
        settled_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          description: |
            When the policy decision expires and the Action must be
            re-evaluated (e.g. balance staleness window).
        policy_version:
          type: integer
          description: Policy version the decision was evaluated against.
        matched_rule:
          type: string
          nullable: true
          description: Rule id that produced the decision.
        reason:
          type: object
          nullable: true
          description: |
            Structured "why" payload when `decision !== "ALLOW"`. Schema
            is decision-dependent — e.g. for DENY:
            `{ code: "counterparty_unverified", counterparty_id: "..." }`.
          additionalProperties: true
        signed_verdict:
          type: string
          nullable: true
          description: |
            EIP-712 signed PolicyDecision attestation. Present when the
            policy was registered on-chain. Opaque to the SDK in v0.3;
            schema formalizes in a later release.
        audit_events:
          type: array
          description: |
            Audit event ids touching this action, ordered by
            `created_at` ascending. Use `brain.audit.get(eventId)` or
            `brain.audit.proof(eventId)` to dereference.
          items: { type: string }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...